refactor(state): extract chain to chain.rs
Move Chain and ChainIterator to dedicated chain.rs module.
