+++ src/state/mod.rs
@@ -1,5 +1,7 @@
+pub mod chain;
 pub mod chainstate;
 pub mod context;
 
+pub use chain::{Chain, ChainIterator};
 pub use chainstate::{ChainstateManager, ChainstateManagerOptions};
 pub use context::{ChainParams, ChainType, Context, ContextBuilder};
